+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change Log
 
 #### Next release
+- Changed Jenkins logo on monitor
 - Improved rebuild plugin integration
 
 #### v0.4.0-build.1565952664 ~ 29.08.16